+++ src/base/kernel/Base.cpp
@@ -176,12 +176,8 @@ int xmrig::Base::init()
 
     Platform::init(config()->userAgent());
 
-#   ifndef XMRIG_PROXY_PROJECT
-    Platform::setProcessPriority(config()->cpu().priority());
-#   endif
-
     if (isBackground()) {
-        Log::background = true;
+        Log::setBackground(true);
     }
     else {
         Log::add(new ConsoleLog());