+++ src/base/kernel/config/BaseConfig.h
@@ -31,9 +31,6 @@
 #include "base/net/stratum/Pools.h"
 
 
-struct option;
-
-
 namespace xmrig {
 
 
@@ -43,7 +40,7 @@ class IJsonReader;
 class BaseConfig : public IConfig
 {
 public:
-    BaseConfig();
+    BaseConfig() = default;
 
     inline bool isAutoSave() const                 { return m_autoSave; }
     inline bool isBackground() const               { return m_background; }
@@ -85,6 +82,8 @@ protected:
 
 private:
     inline void setPrintTime(uint32_t printTime) { if (printTime <= 3600) { m_printTime = printTime; } }
+
+    void setVerbose(const rapidjson::Value &value);
 };