+++ src/base/kernel/interfaces/IConfig.h
@@ -89,8 +89,12 @@ public:
         AssemblyKey          = 1015,
         RandomXInitKey       = 1022,
         RandomXNumaKey       = 1023,
+        RandomXModeKey       = 1029,
+        RandomX1GbPagesKey   = 1031,
+        RandomXWrmsrKey      = 1032,
         CPUMaxThreadsKey     = 1026,
         MemoryPoolKey        = 1027,
+        YieldKey             = 1030,
 
         // xmrig amd
         OclPlatformKey       = 1400,
@@ -125,6 +129,7 @@ public:
         TlsProtocolsKey      = 1114,
         AlgoExtKey           = 1115,
         ProxyPasswordKey     = 1116,
+        LoginFileKey         = 'L',
 
         // xmrig nvidia
         CudaMaxThreadsKey    = 1200,