+++ src/core/config/ConfigTransform.cpp
@@ -37,9 +37,7 @@ static const char *kTls  = "tls";
 }
 
 
-xmrig::ConfigTransform::ConfigTransform()
-{
-}
+xmrig::ConfigTransform::ConfigTransform() = default;
 
 
 void xmrig::ConfigTransform::transform(rapidjson::Document &doc, int key, const char *arg)
@@ -68,7 +66,6 @@ void xmrig::ConfigTransform::transform(rapidjson::Document &doc, int key, const 
     case IConfig::ProxyPasswordKey: /* --access-password */
         return set(doc, "access-password", arg);
 
-    case IConfig::VerboseKey: /* --verbose */
     case IConfig::DebugKey:   /* --debug */
         return transformBoolean(doc, key, true);
 
@@ -112,9 +109,6 @@ void xmrig::ConfigTransform::transform(rapidjson::Document &doc, int key, const 
 void xmrig::ConfigTransform::transformBoolean(rapidjson::Document &doc, int key, bool enable)
 {
     switch (key) {
-    case IConfig::VerboseKey: /* --verbose */
-        return set(doc, "verbose", enable);
-
     case IConfig::DebugKey: /* --debug */
         return set(doc, "debug", enable);